minor struture changes

+++ DirectX12Engine/Window/Window.cpp
@@ -179,10 +179,10 @@ LRESULT Window::WndProc(HWND hWnd, UINT msg, WPARAM wParam, LPARAM lParam)
 		if (wParam == VK_ESCAPE) {
 			DestroyWindow(this->m_hwnd);
 		}
-		Input::PressKey(wParam);		
+		Input::PressKey(static_cast<UINT>(wParam));		
 		return 0;
 	case WM_KEYUP:
-		Input::UnPressKey(wParam);
+		Input::UnPressKey(static_cast<UINT>(wParam));
 		return 0;
 		
 	case WM_DESTROY:
